Align consecutive declarations
Mostly this is to get alignment in the fields within a struct.  It can't
be restricted to solely struct/union/class field declarations, however,
so it will affect all consecutive declarations.  Since this code
originates from C there are lots of consecutive declarations at the
start of a block, which could lead to some formatting churn.  However,
as declarations are localize to the point of first use, the churn should
lessen.