Revert DATASRC::name to char*
Because this data structure is allocated with malloc and not new, the
constructor for the std::string wasn't being run.  Instead of manually
invoking the constructor on the allocated memory, just revert back to
using a C style string for the name field.  Later when the homebrew list
allocation is replaced with std::vector, the fields of these list
nodes can be migrated to classes.