Decouple tests from getenv
Trying to unset environment variables is unreliable between platforms;
sometimes the attempt to unset the environment variable results in
getenv() returning an empty string value instead of nullptr.  This
confuses logic that attempts to determine if a variable is set.  To
resolve this environment variales are loaded through an indirect
function that defaults to getenv but is overridden by tests.  This
allows the tests to more strictly control the results of queries to the
environment.